移除字段、Bug修复

See merge request web/myopia-pac/myopia-pac-backend!655

+++ myopia-business/business-core/school-core/src/main/java/com/wupol/myopia/business/core/school/domain/dto/StudentExportDTO.java
@@ -92,7 +92,7 @@ public class StudentExportDTO implements Serializable {
     private Integer visitsCount;
 
     @ExcelProperty({TOP_HEADER, "问卷次数"})
-    private Integer questionCount;
+    private Long questionCount;
 
     @ExcelProperty({TOP_HEADER, "最新筛查日期"})
     private String lastScreeningTime;
